requirements.txt
aiogram==3.25.0
chempy
python-dotenv==1.2.1